pyproject.toml
[project]
name = "borneo-bulletin-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py (>=2.6.4,<3.0.0)",
    "playwright (>=1.56.0,<2.0.0)",
    "pytest-playwright (>=0.7.2,<0.8.0)",
    "schedule (>=1.2.0,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "jsonschema (>=4.0.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=25.11.0,<26.0.0)",
    "isort (>=7.0.0,<8.0.0)"
]

[tool.poetry]
package-mode = false

[tool.ruff]
select = ["E", "F"]
